Stock/programs/創業支援・新規事業開発（AIエージェント）/projects/DentalInstagramScraper/requirements.txt
# Instagram Data Collection
instaloader==4.10.3

# AI/LLM Integration
anthropic>=0.40.0

# Configuration Management
pyyaml==6.0.1
python-dotenv==1.0.0

# Web Scraping & Parsing
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Browser Automation
playwright>=1.40.0

# Progress Display
tqdm==4.66.1

# Testing (optional)
pytest==7.4.4
pytest-cov==4.1.0
